AVRCP: Fix getFolderItems with all items request
start item and end item are uint32_t
0xFFFFFFFF is commonly used as an indicator for "all items" end value.

Using a jint (int32_t) doesn't work, use jlong instead.

Test: browse with audi kit
Bug: 35385700
Bug: 36645066
(cherry picked from commit aabff2d74224e1d122a2ae767330a69b9324779e)

+++ src/com/android/bluetooth/avrcp/AvrcpHelperClasses.java
@@ -36,14 +36,14 @@ class AvrcpCmd {
     /* Helper classes to pass parameters from callbacks to Avrcp handler */
     class FolderItemsCmd {
         byte mScope;
-        int mStartItem;
-        int mEndItem;
+        long mStartItem;
+        long mEndItem;
         byte mNumAttr;
         int[] mAttrIDs;
         public byte[] mAddress;
 
-        public FolderItemsCmd(byte[] address,byte scope, int startItem, int endItem, byte numAttr,
-                int[] attrIds) {
+        public FolderItemsCmd(byte[] address, byte scope, long startItem, long endItem,
+                byte numAttr, int[] attrIds) {
             mAddress = address;
             this.mScope = scope;
             this.mStartItem = startItem;
